#29 Fixing pet id type

+++ src/main/groovy/com/jos/dem/vetlog/command/PetCommand.groovy
@@ -27,6 +27,6 @@ class PetCommand implements Command {
 
   @NotNull
   @Min(1L)
-  Integer breed
+  Long breed
 
 }
